"make testbuild" and "make testrun" now work, at least for non-MPI builds. scexception.{h,cc} were moved from misc to class to avoid library cross dependency problems.
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%403999

+++ mpqc/src/lib/math/scmat/disttest.cc
@@ -90,7 +90,7 @@ main(int argc, char** argv)
   RefSCDimension d2; d2 << keyval->describedclassvalue("d2");
   RefSCDimension d3; d3 << keyval->describedclassvalue("d3");
 
-  int nblocks = (int)sqrt(msg->n());
+  int nblocks = (int)sqrt(double(msg->n()));
 
   // replace dimensions with dimensions that have subblocks
   d1 = new SCDimension(d1.n(), nblocks);
